v3 iteration of grammar correction prompt.
